Refactor CNI packages
