+++ pkg/cni/cilium.go
@@ -14,13 +14,15 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package kubermatic
+package cni
 
 import (
 	"encoding/json"
 	"fmt"
 
 	appskubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/apps.kubermatic/v1"
+	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
+	"k8c.io/kubermatic/v2/pkg/resources"
 	"k8c.io/kubermatic/v2/pkg/resources/reconciling"
 
 	"k8s.io/apimachinery/pkg/runtime"
@@ -31,7 +33,9 @@ const (
 	ciliumHelmChartName = "cilium"
 )
 
-func CiliumCNIApplicationDefinitionCreator() reconciling.NamedAppsKubermaticV1ApplicationDefinitionCreatorGetter {
+// CiliumApplicationDefinitionCreator creates Cilium ApplicationDefinition managed by KKP to be used
+// for installing Cilium CNI into KKP usr clusters.
+func CiliumApplicationDefinitionCreator() reconciling.NamedAppsKubermaticV1ApplicationDefinitionCreatorGetter {
 	return func() (string, reconciling.AppsKubermaticV1ApplicationDefinitionCreator) {
 		return ciliumAppDefName, func(app *appskubermaticv1.ApplicationDefinition) (*appskubermaticv1.ApplicationDefinition, error) {
 			app.Labels = map[string]string{
@@ -88,4 +92,45 @@ func CiliumCNIApplicationDefinitionCreator() reconciling.NamedAppsKubermaticV1Ap
 			return app, nil
 		}
 	}
+}
+
+// GetCiliumAppInstallOverrideValues returns Helm values to be enforced on the cluster's ApplicationInstallation
+// of the Cilium CNI managed by KKP.
+func GetCiliumAppInstallOverrideValues(cluster *kubermaticv1.Cluster, overwriteRegistry string) map[string]interface{} {
+	values := map[string]interface{}{
+		"cni": map[string]interface{}{
+			// we run Cilium as non-exclusive CNI to allow for Multus use-cases
+			"exclusive": "false",
+		},
+		"operator": map[string]interface{}{
+			"securityContext": map[string]interface{}{
+				"seccompProfile": map[string]interface{}{
+					"type": "RuntimeDefault",
+				},
+			},
+		},
+	}
+
+	if cluster.Spec.ClusterNetwork.ProxyMode == resources.EBPFProxyMode {
+		values["kubeProxyReplacement"] = "strict"
+		values["k8sServiceHost"] = cluster.Status.Address.ExternalName
+		values["k8sServicePort"] = fmt.Sprintf("%d", cluster.Status.Address.Port)
+	} else {
+		values["kubeProxyReplacement"] = "disabled"
+	}
+
+	ipamOperator := map[string]interface{}{
+		"clusterPoolIPv4PodCIDR":  cluster.Spec.ClusterNetwork.Pods.GetIPv4CIDR(),
+		"clusterPoolIPv4MaskSize": fmt.Sprintf("%d", *cluster.Spec.ClusterNetwork.NodeCIDRMaskSizeIPv4),
+	}
+	if cluster.IsDualStack() {
+		values["ipv6"] = map[string]interface{}{"enabled": "true"}
+		ipamOperator["clusterPoolIPv6PodCIDR"] = cluster.Spec.ClusterNetwork.Pods.GetIPv6CIDR()
+		ipamOperator["clusterPoolIPv6MaskSize"] = fmt.Sprintf("%d", *cluster.Spec.ClusterNetwork.NodeCIDRMaskSizeIPv6)
+	}
+	values["ipam"] = map[string]interface{}{"operator": ipamOperator}
+
+	// TODO (rastislavs): override *.image.repository + set *.image.useDigest=false if registry override is configured
+
+	return values
 }
