+++ pkg/cni/doc.go
@@ -15,9 +15,7 @@ limitations under the License.
 */
 
 /*
-Package cni contains a controller that watches Cluster resources,
-and if the CNI for the Cluster is manged by the Applications infra,
-reconciles ApplicationInstallation Resources in the user cluster
-with necessary CNI configuration in ApplicationInstallation's Values.
+Package cni contains supported sCNI version definitions and helpers for managing
+CNIs in KKP.
 */
 package cni
