+++ pkg/controller/seed-controller-manager/cni-application-installation-controller/OWNERS
@@ -1,0 +1,13 @@
+# See the OWNERS docs: https://git.k8s.io/community/contributors/guide/owners.md
+
+approvers:
+  - sig-networking
+
+reviewers:
+  - sig-networking
+
+labels:
+  - sig/networking
+
+options:
+  no_parent_owners: true
