+++ pkg/controller/seed-controller-manager/cni-application-installation-controller/doc.go
@@ -1,0 +1,23 @@
+/*
+Copyright 2022 The Kubermatic Kubernetes Platform contributors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+/*
+Package cniapplicationinstallationcontroller contains a controller that watches
+Cluster resources, and if the CNI for the Cluster is manged by the Applications infra,
+reconciles ApplicationInstallation Resources in the user cluster
+with necessary CNI configuration in ApplicationInstallation's Values.
+*/
+package cniapplicationinstallationcontroller
